[wip] Start of APIs for data and web. OMAService will be the general internal 'public api' that can be used throughout the app. OMADataService will be an unopinionated YSIWYG API that does what you tell it to. - checking must be done before it reaches this, else face the problems you told it to make. API can be used directly for specific instructions and raw data. MVC will then be used for the app we make with it.
May consider better handling of invalid cases to provide more detailed reporting.
- if it matters at some point.

+++ Models/Internal/Score.cs
@@ -7,5 +7,5 @@ class Score
     public float Accuracy { get; set; }
     public long MaxCombo { get; set; }
     public bool PerfectCombo { get; set; }
-    public int PP { get; set; }
+    public float PP { get; set; }
 }
